relocate stories to respective components and removing unused utils file
