+++ src/components/TimePicker24Hour/TimePicker24Hour.stories.jsx
@@ -1,19 +1,19 @@
 import React from "react";
-import TimePickerCarbon from "../components/TimePicker/TimePicker.jsx";
 import { IntlProvider } from "react-intl";
 import moment from "moment";
+import TimePicker24Hour from "./TimePicker24Hour.jsx";
 
 export default {
-  title: "Time Picker with AM and PM",
+  title: "Time Picker 24 hour",
 };
 
 export const Primary = () => {
   return (
     <IntlProvider locale={"en"}>
-      <TimePickerCarbon
+      <TimePicker24Hour
+        onChange={() => {}}
         labelText={"Start Time"}
         isRequired={true}
-        onChange={() => {}}
         translationKey={"APPOINTMENT_TIME_FROM_LABEL"}
         defaultTranslationKey={"Start Time"}
         defaultTime={moment()}
